+++ src/agent_harness/apps/template/__init__.py
@@ -1,0 +1,1 @@
+"""Template app for new vertical scenarios."""